move VAE to convolutional_vae.py and let it work with Variational()
